backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
open-interpreter
pytest
httpx
fastapi-structlog
asgi-correlation-id